requirements.txt
# Pending to move to pyproject
fastapi

uvicorn

sqlalchemy

psycopg2-binary

beautifulsoup4

requests

sentence-transformers

torch==2.8.0

pgvector

dashscope

# Para clustering y análisis
scikit-learn

numpy

python-dotenv